refactor(encoding): move match-generator tests to a tests submodule
Relocate the ~4.9k lines of unit / round-trip tests (parse x search
matrix, per-level parameter resolution, greedy/lazy/optimal round-trips,
dictionary priming) out of the driver module body into a dedicated
`tests` submodule file. Production code stays in mod.rs; the test file is
`#[cfg(test)] mod tests;`.

match_generator/mod.rs: 7000 -> 2063 (production only).

Part of #441.
